# api/swagger.yaml
...
swagger: "2.0"
info:
  version: "1.0"
  title: "ExpressionService"
  contact: {}
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /createUser:
    post:
      tags:
      - "ExpressionPortBinding"
      summary: "createUser"
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/createUser"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/createUserResponse"
      x-swagger-router-controller: "ExpressionPortBinding"
...
  /deleteExpression:
    post:
      tags:
      - "ExpressionPortBinding"
      summary: "deleteExpression"
      operationId: "deleteExpression"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "Body"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/deleteExpression"
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/deleteExpressionResponse"
      x-swagger-router-controller: "ExpressionPortBinding"
...
definitions:
...
  createUserResponse:
    type: "object"
    required:
    - "parameters"
    properties:
      parameters:
        type: "string"
    title: "createUserResponse"
    example:
      parameters: "parameters"
...
  deleteExpressionResponse:
    type: "object"
    required:
    - "parameters"
    properties:
      parameters:
        type: "string"
    title: "deleteExpressionResponse"
    example:
      parameters: "parameters"
...
  createUser:
    type: "object"
    required:
    - "parameters"
    properties:
      parameters:
        type: "string"
    title: "createUser"
...
  deleteExpression:
    type: "object"
    required:
    - "parameters"
    properties:
      parameters:
        type: "string"
    title: "deleteExpression"
...